Dart: Refactors Proxies
This change implements some of the good suggestions from:

https://github.com/domokit/mojo/issues/647

In particular, it does away with the wierd 'ptr' thing, is more
careful to mangle names that could conflict, and hopefully has
an easier-to-follow structure. This comes at the cost of needing
an indirection to access "control" functions on a Proxy, as
suggested by floitsch@ (though I have punched holes for close()
and responseOrError() since they show up a lot).

related #647

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1964193002 .