Dart: Refactor Stubs
This change refactors Stubs to be more similar to Proxies.
In particular, after this change, both FooStub and FooProxy
will implement the Foo interface.

After this change there are also more opportunities for
Stubs and Proxies to share interfaces, which will allow
both client and internal code to avoid using Object and
dynamic, and I'll pursue those in a follow up CL.

As with the change for Proxies, this is a breaking change
that will require indirection through the .ctrl field of
a Stub to manipulate the underlying MojoMessageHandler.
There are no other user-visible changes.

related #647

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1983453002 .