Mojom frontend: Cycle detection: Don't suppress simple type fields in unions.
When modelling the type graph we were making an "optimization"
by ignoring fields of simple type. That is, rather than adding an
edge to a leaf node for a field of simple type, we were not adding an
edge at all. This is correct for structs which are modeled as regular
circle nodes but it is not correct for unions which are modeled as square
nodes. In this patch we fix this by keeping track of the fact that there
was at least one field of this type and adding an edge to a synthetic
leaf node.

I cannot explain why the .mojom.dart files needed to be re-generated. This patch
should not have affected them. It must have been some earlier patch.

BUG= fixes #793
R=azani@chromium.org

Review URL: https://codereview.chromium.org/2045993005 .