MOTOWN: Fix the demand model for A/V scenarios This CL addresses an issue with the new mojo packet transport demand model implementation and the way it translates to the in-proc streaming framework's positive/neutral/negative demand model. Specifically, the producer was translating lack of mojo demand into neutral framework demand. This meant that, when the demux had to generate a bnuch of undemanded video packets to find a demanded audio packet, the video packets flowed downstream and got decoded prematurely. This eats up a lot of memory, because uncompressed video packets are so large.
Translating lack of mojo demand into negative framework demand has the
desired effect of keeping undemanded packets upstream and undecoded.
However, the demux (which is hosted by ActiveMultistreamSourceStage) was
designed to queue only one packet, which meant that negative demand on
one stream could block the flow of packets on a stream where there's
positive demand. As part of this fix, ActiveMultistreamSourceStage had
to be changed to queue an arbitrary number of packets in per-stream
queues. That allows packets to queue up on a stream with negative demand
while packets flow freely on streams with positive demand.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/2189823003 .