MOTOWN: Add SharedBufferSet for dynamic buffer allocation, consumer end The mojo transport definition was recently changed to allow multiple shared buffers to be used per connection with buffers dynamically added and removed. So far, this capability has not been used.
This CL introduces the class SharedBufferSet, which maintains a set of
MappedSharedBuffers and knows how to translate addresses into buffer_id/
offset pairs (aka 'Locator's) and back. This is what the consumer will
use, and it will serve as a base class for the allocator on the producer
end of the connection.

The CL includes a unit test.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/2202693003 .