MOTOWN: Stop propagating Prime through the graph...it's a renderer thing One of the consequences of the new flow control design is that priming the graph can be done at the renderers. The renderers adjust their demand, which ripples up the graph. The old priming logic started at the demux and traversed the graph to the renderers. Flush still needs to start at the demux.
Prime is now a method on MediaTimelineControlPoint, and the media
timeline controller is in charge of distributing the signal to the
renderers.

Included is a TODO for adding a rate to the Prime method. This will
allow renderers to e.g. queue up twice as much data when the intended
rate is 2x.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/2219533003 .