MOTOWN: Asynchronous problem reporting Prior to this CL, players and their subordinates had no way to report problems to the client/user. This CL addresses that problem by defining a representation for problems (Problem mojom struct) and including this representation in various component status structures.
Problems are roughly defined as issues that prevent a component from
behaving as the client intends them to.

This scheme for representing problems is a bit unconventional, but works
very well for loosely-coupled async interfaces such as player, source,
demux, etc. The component can report problems at any time via async
status updates. Problems are modelled as part of a component's state,
which is appropriate given that problems persist for some time interval.
There is no general means of dismissing problems...a problem remains
part of the state until the underlying cause is addressed.

The client, which consumes problem structures, needs to be able to
distinguish different types of problems and take the appropriate action.
In most cases, the correct action will be to report the problem to the
user. While it's useful to put all the details of the problem in the
structure to enable this, there's no reason why problem UI can't
interrogate other sources to get needed information. Sometimes problem
UI will provide guidance on how the user might solve the problem. Other
times it might just tell them that things are busted. In some cases, the
client may have its own way of solving specific problems without user
intervention.

The current struct definition for problems is probably inadequate.
In any case, we are currently only using problem reporting for simple
stuff, like "the file you want to play doesn't exist", so it'll do.

This kind of problem reporting is a generally-useful idiom and should be
moved out of media, if we can get consensus to do so.

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/2213283002 .