Change the canonical way to include the C result.h to <mojo/result.h>.
Probably the system-specific stuff should be moved out of that result.h
to one in the (currently-nonexistent) system subdirectory, but we'll
worry about that later.

Also, like for macros.h, I've left the includes under
mojo/public/{c,cpp} alone for now.

And the way I deal with cgo is super-hacky. (To be fair, it was hacky
and wrong before too, just less obviously so.)

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/2224843004 .