Rust: Add encoder/decoder tests
Uses the validation input test parser library in order to automate
running tests on "good" inputs to the mojom encoding and decoding
frameworks in the Rust bindings. Included is Rust code generated by
the Mojom code generator (to be added in a later CL) for
validation_test_interfaces.mojom, which is then used in the tests.

R=eholk@chromium.org

Review URL: https://codereview.chromium.org/2213563002 .