Rust: Add validation checks to decode
In the Rust bindings we now perform validation of the input
buffer on Mojom decode. The generated code checked in for
tests has been updated as well in order to test validation in
nearly all of the cases outlined in the "golden" files.

The only missing tests to run are the integration tests which
are to come in a future CL.

R=eholk@chromium.org

Review URL: https://codereview.chromium.org/2220183003 .