Rust: Add bindings generator
Add the initial implementation of the bindings generator
which produces Rust code that compiles and executes. This
generator only supports the encoding framework in the Rust
bindings.

The bindings generator is not tested directly and is
instead currently tested via the generated code checked in
with the Rust bindings tests (generated from
validation_test_interfaces.mojom).

R=eholk@chromium.org

Review URL: https://codereview.chromium.org/2189263002 .